Changed data types to be integers

+++ db/migrate/20190830212846_add_mau_to_app.rb
@@ -1,9 +1,9 @@
 class AddMauToApp < ActiveRecord::Migration
   def change
-    add_column :android_apps, :mau, :decimal
-    add_column :android_apps, :mau_monthly_change, :decimal
+    add_column :android_apps, :mau, :integer
+    add_column :android_apps, :mau_monthly_change, :integer
 
-    add_column :ios_apps, :mau, :decimal
-    add_column :ios_apps, :mau_monthly_change, :decimal
+    add_column :ios_apps, :mau, :integer
+    add_column :ios_apps, :mau_monthly_change, :integer
   end
 end
